scripts/requirements.txt
../replays_unpack
lxml
GitPython
inflection